fix: move TaskRepository.java to package repository
