+++ src/main/java/ru/mahotin/repository/TaskRepository.java
@@ -1,4 +1,4 @@
-package ru.mahotin;
+package ru.mahotin.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
